+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [[tbd]] (2024-04-22)
+
+### Changed
+- LT-5418: Extend listener registration API with service provider
+
 ## 13.1.0 (2024-04-22)
 
 ### Added